+++ src/test/java/seedu/ccacommander/logic/parser/CcaCommanderParserTest.java
@@ -36,7 +36,7 @@ import seedu.ccacommander.testutil.MemberBuilder;
 import seedu.ccacommander.testutil.MemberUtil;
 
 
-public class AddressBookParserTest {
+public class CcaCommanderParserTest {
 
     private final CcaCommanderParser parser = new CcaCommanderParser();
